Provide multi-language support
Add an I18n mechanism for application translation in multiple languages.

Translations are stored in resources bundles, supporting a flexible organization
of locale files via a fallback mecanism.

The locale to use for the request can be set via a number of mechanisms. It is either
read from the url, implied from a previous session, or inferred via the `Accept-Language`
header.