Display validation errors when authentication fails
Show a list of error messages when authentication fails,
yet we're still using a fake authenticator. Error translations
are read from view properties files, in the user locale.

Provide a way to describe/define error keys as Kotlin code.
Keys are named after the constants and object hierarchies that
hold them, with underscores converted to hyphens.