Separate tests from acceptance tests in build
Add a new source set for acceptance tests so that we can run them
independently using the `acceptanceTest` task. This change makes
possible to improve the Concourse pipeline by separating the test
run from the acceptance test run, which requires a selenium server
to run.

Move feature tests in the server module to the acceptance source set,
adjust gradle dependencies declaration and add make targets for
tests and acceptance tests.