Separate test and acceptance tests jobs in the build pipeline
Currently, the Concourse CI pipeline defines a single build job
which runs all the tests. Yet, acceptance tests are now in a separate
source set in the gradle build so that we can run them independently.

This change splits the pipeline in a test job and a future acceptance
job. The test job only runs the tests, not the acceptance tests.

Note: the test job is based on the official gradle image, it does
not run using the gradle wrapper.